+++ src/BA-Trees/main.jl
@@ -17,7 +17,7 @@ Funzione 'principale' che fa uso di run_function e trad_function
 definite negli altri file.
 """
 function batrees(f; dataset_name="iris", num_trees=5, max_depth=3)
-    WRAP_batrees(f; dataset_name="iris", num_trees=5, max_depth=3)
+    WRAP_batrees(f; dataset_name, num_trees, max_depth)
     trad()
 end
